flask/requirements.txt
certifi==2018.4.16
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
pymongo==3.6.1
Werkzeug==0.14.1
google-cloud==0.34.0